completing the database
